requirements.txt
attrs==23.2.0
blinker==1.7.0
branca==0.7.0
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
fiona==1.9.5
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
folium==0.15.1
fonttools==4.47.2
geopandas==0.14.2
greenlet==3.0.3
gunicorn==22.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mplleaflet==0.0.5
mysql-connector-python==8.3.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pandas==2.1.4
pillow==10.2.0
plotly==5.18.0
pyparsing==3.1.1
pypng==0.20220715.0
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2023.3.post1
qrcode==7.4.2
reportlab==4.2.0
requests==2.31.0
segno==1.6.1
shapely==2.0.2
six==1.16.0
SQLAlchemy==2.0.25
tenacity==8.2.3
typing_extensions==4.10.0
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
xyzservices==2023.10.1
